requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
ccxt
pandas
numpy
python-dotenv
requests
websocket-client
stripe
Werkzeug
gunicorn
psycopg2-binary